requirements.txt
alembic==1.14.1
aniso8601==10.0.1
asttokens==3.0.0
backcall==0.2.0
bcrypt==4.3.0
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cors==1.0.1
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
executing==2.2.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
limits==3.13.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2024.4.1
platformdirs==4.3.6
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.41
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
tldextract==5.1.2
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.2.3
virtualenv==20.31.0
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
